+++ components/script/dom/node/mod.rs
@@ -7,8 +7,7 @@ pub(crate) mod children_mutation;
 pub(crate) mod context;
 pub(crate) mod focus;
 pub(crate) mod iterators;
-pub(crate) mod layout;
-pub(crate) mod methods;
+pub(crate) mod layout_dom;
 #[allow(clippy::module_inception, reason = "The interface name is node")]
 pub(crate) mod node;
 pub(crate) mod nodeiterator;
