+++ components/script/layout_dom/servo_layout_node.rs
@@ -26,7 +26,7 @@ use style::selector_parser::PseudoElement;
 use super::ServoLayoutElement;
 use crate::dom::bindings::root::LayoutDom;
 use crate::dom::element::Element;
-use crate::dom::layout::NodeTypeIdWrapper;
+use crate::dom::layout_dom::NodeTypeIdWrapper;
 use crate::dom::node::{Node, NodeFlags};
 use crate::layout_dom::{
     ServoDangerousStyleNode, ServoLayoutDomTypeBundle, ServoLayoutNodeChildrenIterator,
